+++ tests/test_custom_types.py
@@ -1,5 +1,6 @@
+import sys
 from enum import Enum as BuiltinEnum, auto
-from typing import Optional, Union
+from typing import List, Optional, Tuple, TypeVar, Union
 
 from typing_extensions import Literal
 
@@ -7,10 +8,14 @@ from argtyped import Choices, Enum as ArgtypedEnum
 from argtyped.custom_types import (
     is_choices,
     is_enum,
+    is_list,
     is_optional,
     unwrap_choices,
+    unwrap_list,
     unwrap_optional,
 )
+
+T = TypeVar("T")
 
 
 def test_is_enum():
@@ -63,3 +68,31 @@ def test_unwrap_choices():
     assert unwrap_choices(Choices["a", "b", "c"]) == ("a", "b", "c")
     assert unwrap_choices(Choices[["a", "b"] + ["c"]]) == ("a", "b", "c")
     assert unwrap_choices(Literal["a", "b", "c"]) == ("a", "b", "c")
+
+
+def test_is_list():
+    assert is_list(List[int])
+    assert is_list(List[T][int])
+    assert is_list(List[Optional[int]])
+    assert is_list(List[List[int]])
+    assert is_list(List[Union[int, float]])
+    assert not is_list(Optional[List[Optional[int]]])
+    assert not is_list(Tuple[int, int])
+    if sys.version_info >= (3, 9):
+        # PEP-585 was implemented in Python 3.9, which allows using standard types as
+        # typing containers.
+        assert is_list(list[int])
+        assert is_list(list[T][int])
+        assert not is_list(Optional[list[int]])
+        assert not is_list(dict[str, list[int]])
+        assert not is_list(tuple[int, int])
+
+
+def test_unwrap_list():
+    assert unwrap_list(List[int]) == int
+    assert unwrap_list(List[T][int]) == int
+    assert unwrap_optional(unwrap_list(List[Optional[int]])) == int
+    if sys.version_info >= (3, 9):
+        assert unwrap_list(list[int]) == int
+        assert unwrap_list(list[T][int]) == int
+        assert unwrap_optional(unwrap_list(list[Optional[int]])) == int